+++ crates/fluvio-spu/src/services/public/tests/produce.rs
@@ -1,0 +1,1 @@
+
