+++ crates/fluvio-spu/src/services/public/tests/stream_fetch.rs
@@ -1,11 +1,12 @@
 use std::{
     env::temp_dir,
-    path::{Path, PathBuf},
+    path::{PathBuf, Path},
     sync::atomic::{AtomicU16, Ordering},
     time::Duration,
 };
 
 use flate2::{Compression, bufread::GzEncoder};
+
 use fluvio_controlplane_metadata::{
     partition::Replica,
     smartmodule::{SmartModule, SmartModuleWasm, SmartModuleWasmFormat},
@@ -30,7 +31,7 @@ use fluvio_spu_schema::server::{
 use fluvio_spu_schema::server::stream_fetch::SmartModuleWasmCompressed;
 use fluvio_spu_schema::server::stream_fetch::LegacySmartModulePayload;
 use fluvio_spu_schema::server::stream_fetch::{DefaultStreamFetchRequest};
-use crate::core::GlobalContext;
+use crate::{core::GlobalContext, services::public::tests::create_filter_records};
 use crate::config::SpuConfig;
 use crate::replication::leader::LeaderReplicaState;
 use crate::services::public::create_public_server;
@@ -46,6 +47,45 @@ use dataplane::{
 };
 
 static NEXT_PORT: AtomicU16 = AtomicU16::new(12000);
+
+fn read_filter_from_path(filter_path: impl AsRef<Path>) -> Vec<u8> {
+    let path = filter_path.as_ref();
+    std::fs::read(path).unwrap_or_else(|_| panic!("Unable to read file {}", path.display()))
+}
+
+fn zip(raw_buffer: Vec<u8>) -> Vec<u8> {
+    use std::io::Read;
+    let mut encoder = GzEncoder::new(raw_buffer.as_slice(), Compression::default());
+    let mut buffer = Vec::with_capacity(raw_buffer.len());
+    encoder
+        .read_to_end(&mut buffer)
+        .unwrap_or_else(|_| panic!("Unable to gzip file"));
+    buffer
+}
+
+fn read_wasm_module(module_name: &str) -> Vec<u8> {
+    let spu_dir = std::env::var("CARGO_MANIFEST_DIR").expect("target");
+    let wasm_path = PathBuf::from(spu_dir)
+        .parent()
+        .expect("parent")
+        .join(format!(
+            "fluvio-smartmodule/examples/target/wasm32-unknown-unknown/release/{}.wasm",
+            module_name
+        ));
+    read_filter_from_path(wasm_path)
+}
+
+fn load_wasm_module<S: ReplicaStorage>(ctx: &GlobalContext<S>, module_name: &str) {
+    let wasm = zip(read_wasm_module(module_name));
+    ctx.smartmodule_localstore().insert(SmartModule {
+        name: module_name.to_owned(),
+        wasm: SmartModuleWasm {
+            format: SmartModuleWasmFormat::Binary,
+            payload: wasm,
+        },
+        ..Default::default()
+    });
+}
 
 #[fluvio_future::test(ignore)]
 async fn test_stream_fetch_basic() {
@@ -195,65 +235,6 @@ async fn test_stream_fetch_basic() {
 
     server_end_event.notify();
     debug!("terminated controller");
-}
-
-/// create records that can be filtered
-fn create_filter_records(records: u16) -> RecordSet {
-    BatchProducer::builder()
-        .records(records)
-        .record_generator(Arc::new(generate_record))
-        .build()
-        .expect("batch")
-        .records()
-}
-
-fn generate_record(record_index: usize, _producer: &BatchProducer) -> Record {
-    let msg = match record_index {
-        0 => "b".repeat(100),
-        1 => "a".repeat(100),
-        _ => "z".repeat(100),
-    };
-
-    Record::new(RecordData::from(msg))
-}
-
-fn read_filter_from_path(filter_path: impl AsRef<Path>) -> Vec<u8> {
-    let path = filter_path.as_ref();
-    std::fs::read(path).unwrap_or_else(|_| panic!("Unable to read file {}", path.display()))
-}
-
-fn zip(raw_buffer: Vec<u8>) -> Vec<u8> {
-    use std::io::Read;
-    let mut encoder = GzEncoder::new(raw_buffer.as_slice(), Compression::default());
-    let mut buffer = Vec::with_capacity(raw_buffer.len());
-    encoder
-        .read_to_end(&mut buffer)
-        .unwrap_or_else(|_| panic!("Unable to gzip file"));
-    buffer
-}
-
-fn read_wasm_module(module_name: &str) -> Vec<u8> {
-    let spu_dir = std::env::var("CARGO_MANIFEST_DIR").expect("target");
-    let wasm_path = PathBuf::from(spu_dir)
-        .parent()
-        .expect("parent")
-        .join(format!(
-            "fluvio-smartmodule/examples/target/wasm32-unknown-unknown/release/{}.wasm",
-            module_name
-        ));
-    read_filter_from_path(wasm_path)
-}
-
-fn load_wasm_module<S: ReplicaStorage>(ctx: &GlobalContext<S>, module_name: &str) {
-    let wasm = zip(read_wasm_module(module_name));
-    ctx.smartmodule_localstore().insert(SmartModule {
-        name: module_name.to_owned(),
-        wasm: SmartModuleWasm {
-            format: SmartModuleWasmFormat::Binary,
-            payload: wasm,
-        },
-        ..Default::default()
-    });
 }
 
 async fn legacy_test<Fut, TestFn>(
