+++ crates/fluvio-spu/src/services/public/tests/test_data/filter_missing_attribute.wasm
